+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - [SCaddins] better tooltip help added to ribbon.
 - [SCexport] If destination file exists -> check if it's open/available before attempting an export.
 - [SCexport] removed non-print line-type feature. (causing problems with permissions)
+- [SCexport] correctly hide title blocks in when exporting from Revit 2015.
 - [SCexport] check for illegal file names and replace invalid chars with a '_'
 - **[SCexport] correctly assign and save AutoCAD export version.**